Bumped versions to 8.5 Replaced obsolete stuff with NotImplementedException Bumped counts to long Compiles...

+++ Our.Umbraco.ContentList/DataSources/Listables/IListableDataSource.cs
@@ -5,6 +5,6 @@ namespace Our.Umbraco.ContentList.DataSources.Listables
     public interface IListableDataSource
     {
         IQueryable<IListableContent> Query(PagingParameter pagingParameter);
-        int Count(int preSkip);
+        long Count(long preSkip);
     }
 }
